Fixed next and get started  button Size correction
